Теперь аудиоменеджер работает в режиме синглтон
