+++ Assets/Resources/Prefabs/UiPrefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 53be7bdd3c7834033b063b9c8b3c14de
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
