+++ Assets/Resources/Prefabs/UiPrefabs/SettingsMenu.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a0f559117ef514874a6b1ecda193e547
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
