+++ Assets/Scripts/Audio/AudioManager.cs
@@ -4,91 +4,32 @@ using UnityEngine;
 using UnityEngine.Audio;
 
 
-[RequireComponent(typeof(AudioSource))]
 public class AudioManager : MonoBehaviour
 {
+    public static AudioManager Instance;
     public const string MasterVolume = "Master";
     public const string MusicVolume = "Music";
     public const string SFXVolume = "SFX";
 
-    [SerializeField] private AudioMixer _audioMixer;
-
-    [SerializeField] private List<AudioClip> _trackList;
-    private int _currentTrackIndex;
-
-    private AudioSource _audioPlayer;
-    private GameObject audioManager;
-
+    public AudioMixer _audioMixer;
     private void Awake()
     {
-        _audioPlayer = GetComponent<AudioSource>();
-        GameObject audioManager = GameObject.FindGameObjectWithTag("AudioManager");
-        DontDestroyOnLoad(audioManager);
-    }
-
-    private void Start()
-    {
-        Initialize();
-        PlayNextTrack();
-    }
-
-    private void Initialize()
-    {
-        ShuffleTracks();
-        _currentTrackIndex = 0;
-        _audioPlayer.pitch = 1;
-    }
-
-    private void PlayNextTrack()
-    {
-        _audioPlayer.clip = _trackList[_currentTrackIndex];
-        _audioPlayer.Play();
-        StartCoroutine(PlayNextTrackAfterDelay(_audioPlayer.clip.length));
-    }
-
-    private IEnumerator PlayNextTrackAfterDelay(float delay)
-    {
-        yield return new WaitForSeconds(delay);
-        if (_currentTrackIndex >= _trackList.Count - 1)
+        if (Instance != null)
         {
-            ShuffleTracks();
-            _currentTrackIndex = 0;
+            Destroy(gameObject);
         }
         else
         {
-            _currentTrackIndex++;
+            Instance = this;
         }
-        PlayNextTrack();
+
+        DontDestroyOnLoad(Instance.gameObject);
     }
 
-    private void ShuffleTracks()
-    {
-        System.Random random = new();
-        for (int n = _trackList.Count - 1; n > 1; n--)
-        {
-            int k = random.Next(n + 1);
-            (_trackList[n], _trackList[k]) = (_trackList[k], _trackList[n]);
-        }
-    }
-
-    public void OnMasterVolumeChanged(System.Single masterLevel)
-    {
-        SetMixerValue(MasterVolume, masterLevel);
-    }
-
-    public void OnMusicVolumeChanged(System.Single musicLevel)
-    {
-        SetMixerValue(MusicVolume, musicLevel);
-    }
-
-    public void OnSFXVolumeChanged(System.Single sfxLevel)
-    {
-        SetMixerValue(SFXVolume, sfxLevel);
-    }
-
-    private void SetMixerValue(string key, System.Single value)
+    public void SetMixerValue(string key, float value)
     {
         value = value == 0 ? -80 : Mathf.Log10(value) * 20;
+        // Debug.Log($"Set {key} volume to {value}");
         _audioMixer.SetFloat(key, value);
     }
 }
