+++ Assets/Scripts/Audio/MusicManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e8ac33a43b58e4878abe0e033266bf9d
+guid: 5bbf4afcc86ab48b9be9e754f01287a5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
