+++ Assets/Scripts/Audio/SliderHandler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e8ac33a43b58e4878abe0e033266bf9d
+guid: bfaf8b77e1bb340518b67d4a657f2194
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
